scripts/requirements.txt
feedparser>=6.0.0
psycopg2-binary>=2.9.0
anthropic>=0.40.0
scrapling[all]>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=4.0.0
